+++ src/gateway/mod.rs
@@ -1,39 +1,31 @@
-pub mod gateway;
-pub mod handle;
-pub mod heartbeat;
-pub mod message;
+pub mod default;
 
-pub use gateway::*;
-pub use handle::*;
-use heartbeat::*;
-pub use message::*;
+use self::event::Events;
+use crate::types::{
+    self, AutoModerationRule, AutoModerationRuleUpdate, Channel, ChannelCreate, ChannelDelete,
+    ChannelUpdate, Composite, Guild, GuildRoleCreate, GuildRoleUpdate, JsonField, RoleObject,
+    Snowflake, SourceUrlField, ThreadUpdate, UpdateMessage, WebSocketEvent,
+};
+use default::heartbeat::HeartbeatThreadCommunication;
 use tokio_tungstenite::tungstenite::Message;
 
 use crate::errors::GatewayError;
-use crate::types::{Snowflake, WebSocketEvent};
 
 use async_trait::async_trait;
 use std::any::Any;
 use std::collections::HashMap;
-use std::fmt::Debug;
 use std::sync::{Arc, RwLock};
 use std::time::Duration;
-use tokio::time::sleep_until;
 
+pub use default::*;
 use futures_util::stream::SplitSink;
-use futures_util::stream::SplitStream;
-use futures_util::{Sink, StreamExt};
+use futures_util::Sink;
 use futures_util::{SinkExt, Stream};
 use log::{info, trace, warn};
-use tokio::net::TcpStream;
 use tokio::sync::mpsc::Sender;
 use tokio::sync::Mutex;
-use tokio::task;
-use tokio::task::JoinHandle;
-use tokio::time;
-use tokio::time::Instant;
-use tokio_tungstenite::MaybeTlsStream;
-use tokio_tungstenite::{connect_async_tls_with_config, Connector, WebSocketStream};
+
+pub type GatewayStore = Arc<Mutex<HashMap<Snowflake, Arc<RwLock<ObservableObject>>>>>;
 
 // Gateway opcodes
 /// Opcode received when the server dispatches a [crate::types::WebSocketEvent]
@@ -136,53 +128,405 @@ impl<T: WebSocketEvent> GatewayEvent<T> {
     }
 }
 
-#[cfg(test)]
-mod test {
-    use crate::types;
+#[allow(clippy::type_complexity)]
+#[async_trait]
+pub trait GatewayCapable<R, S, G, H>
+where
+    R: Stream,
+    S: Sink<Message>,
+    G: GatewayHandleCapable<R, S>,
+    H: HeartbeatHandlerCapable<S> + Send + Sync,
+{
+    fn get_events(&self) -> Arc<Mutex<Events>>;
+    fn get_websocket_send(&self) -> Arc<Mutex<SplitSink<S, Message>>>;
+    fn get_store(&self) -> GatewayStore;
+    fn get_url(&self) -> String;
+    fn get_heartbeat_handler(&self) -> &H;
+    /// Returns a Result with a matching impl of [`GatewayHandleCapable`], or a [`GatewayError`]
+    ///
+    /// DOCUMENTME: Explain what this method has to do to be a good get_handle() impl, or link to such documentation
+    async fn get_handle(websocket_url: String) -> Result<G, GatewayError>;
+    async fn close(&mut self);
+    /// This handles a message as a websocket event and updates its events along with the events' observers
+    async fn handle_message(&mut self, msg: GatewayMessage) {
+        if msg.is_empty() {
+            return;
+        }
 
-    use super::*;
-    use std::sync::atomic::{AtomicI32, Ordering::Relaxed};
+        if !msg.is_error() && !msg.is_payload() {
+            warn!(
+                "Message unrecognised: {:?}, please open an issue on the chorus github",
+                msg.message.to_string()
+            );
+            return;
+        }
 
-    #[derive(Debug)]
-    struct Consumer {
-        _name: String,
-        events_received: AtomicI32,
-    }
+        if msg.is_error() {
+            let error = msg.error().unwrap();
 
-    #[async_trait]
-    impl Observer<types::GatewayResume> for Consumer {
-        async fn update(&self, _data: &types::GatewayResume) {
-            self.events_received.fetch_add(1, Relaxed);
+            warn!("GW: Received error {:?}, connection will close..", error);
+
+            self.close().await;
+
+            let events = self.get_events();
+            let events = events.lock().await;
+
+            events.error.notify(error).await;
+
+            return;
+        }
+
+        let gateway_payload = msg.payload().unwrap();
+        println!("gateway payload: {:#?}", &gateway_payload);
+
+        // See https://discord.com/developers/docs/topics/opcodes-and-status-codes#gateway-gateway-opcodes
+        match gateway_payload.op_code {
+            // An event was dispatched, we need to look at the gateway event name t
+            GATEWAY_DISPATCH => {
+                let Some(event_name) = gateway_payload.event_name else {
+                    warn!("Gateway dispatch op without event_name");
+                    return;
+                };
+
+                trace!("Gateway: Received {event_name}");
+
+                macro_rules! handle {
+                    ($($name:literal => $($path:ident).+ $( $message_type:ty: $update_type:ty)?),*) => {
+                        match event_name.as_str() {
+                            $($name => {
+                                let events = self.get_events();
+                                let event = &mut events.lock().await.$($path).+;
+                                let json = gateway_payload.event_data.unwrap().get();
+                                match serde_json::from_str(json) {
+                                    Err(err) => warn!("Failed to parse gateway event {event_name} ({err})"),
+                                    Ok(message) => {
+                                        $(
+                                            let mut message: $message_type = message;
+                                            let store = self.get_store();
+                                            let store = store.lock().await;
+                                            let id = if message.id().is_some() {
+                                                message.id().unwrap()
+                                            } else {
+                                                event.notify(message).await;
+                                                return;
+                                            };
+                                            if let Some(to_update) = store.get(&id) {
+                                                let object = to_update.clone();
+                                                let inner_object = object.read().unwrap();
+                                                if let Some(_) = inner_object.downcast_ref::<$update_type>() {
+                                                    let ptr = Arc::into_raw(object.clone());
+                                                    // SAFETY:
+                                                    // - We have just checked that the typeid of the `dyn Any ...` matches that of `T`.
+                                                    // - This operation doesn't read or write any shared data, and thus cannot cause a data race
+                                                    // - The reference count is not being modified
+                                                    let downcasted = unsafe { Arc::from_raw(ptr as *const RwLock<$update_type>).clone() };
+                                                    drop(inner_object);
+                                                    message.set_json(json.to_string());
+                                                    message.set_source_url(self.get_url().clone());
+                                                    message.update(downcasted.clone());
+                                                } else {
+                                                    warn!("Received {} for {}, but it has been observed to be a different type!", $name, id)
+                                                }
+                                            }
+                                        )?
+                                        event.notify(message).await;
+                                    }
+                                }
+                            },)*
+                            "RESUMED" => (),
+                            "SESSIONS_REPLACE" => {
+                                let result: Result<Vec<types::Session>, serde_json::Error> =
+                                    serde_json::from_str(gateway_payload.event_data.unwrap().get());
+                                match result {
+                                    Err(err) => {
+                                        warn!(
+                                            "Failed to parse gateway event {} ({})",
+                                            event_name,
+                                            err
+                                        );
+                                        return;
+                                    }
+                                    Ok(sessions) => {
+                                        let events = self.get_events();
+                                        let events = events.lock().await;
+                                        events.session.replace.notify(
+                                            types::SessionsReplace {sessions}
+                                        ).await;
+                                    }
+                                }
+                            },
+                            _ => {
+                                warn!("Received unrecognized gateway event ({event_name})! Please open an issue on the chorus github so we can implement it");
+                            }
+                        }
+                    };
+                }
+
+                // See https://discord.com/developers/docs/topics/gateway-events#receive-events
+                // "Some" of these are undocumented
+                handle!(
+                    "READY" => session.ready,
+                    "READY_SUPPLEMENTAL" => session.ready_supplemental,
+                    "APPLICATION_COMMAND_PERMISSIONS_UPDATE" => application.command_permissions_update,
+                    "AUTO_MODERATION_RULE_CREATE" =>auto_moderation.rule_create,
+                    "AUTO_MODERATION_RULE_UPDATE" =>auto_moderation.rule_update AutoModerationRuleUpdate: AutoModerationRule,
+                    "AUTO_MODERATION_RULE_DELETE" => auto_moderation.rule_delete,
+                    "AUTO_MODERATION_ACTION_EXECUTION" => auto_moderation.action_execution,
+                    "CHANNEL_CREATE" => channel.create ChannelCreate: Guild,
+                    "CHANNEL_UPDATE" => channel.update ChannelUpdate: Channel,
+                    "CHANNEL_UNREAD_UPDATE" => channel.unread_update,
+                    "CHANNEL_DELETE" => channel.delete ChannelDelete: Guild,
+                    "CHANNEL_PINS_UPDATE" => channel.pins_update,
+                    "CALL_CREATE" => call.create,
+                    "CALL_UPDATE" => call.update,
+                    "CALL_DELETE" => call.delete,
+                    "THREAD_CREATE" => thread.create, // TODO
+                    "THREAD_UPDATE" => thread.update ThreadUpdate: Channel,
+                    "THREAD_DELETE" => thread.delete, // TODO
+                    "THREAD_LIST_SYNC" => thread.list_sync, // TODO
+                    "THREAD_MEMBER_UPDATE" => thread.member_update, // TODO
+                    "THREAD_MEMBERS_UPDATE" => thread.members_update, // TODO
+                    "GUILD_CREATE" => guild.create, // TODO
+                    "GUILD_UPDATE" => guild.update, // TODO
+                    "GUILD_DELETE" => guild.delete, // TODO
+                    "GUILD_AUDIT_LOG_ENTRY_CREATE" => guild.audit_log_entry_create,
+                    "GUILD_BAN_ADD" => guild.ban_add, // TODO
+                    "GUILD_BAN_REMOVE" => guild.ban_remove, // TODO
+                    "GUILD_EMOJIS_UPDATE" => guild.emojis_update, // TODO
+                    "GUILD_STICKERS_UPDATE" => guild.stickers_update, // TODO
+                    "GUILD_INTEGRATIONS_UPDATE" => guild.integrations_update,
+                    "GUILD_MEMBER_ADD" => guild.member_add,
+                    "GUILD_MEMBER_REMOVE" => guild.member_remove,
+                    "GUILD_MEMBER_UPDATE" => guild.member_update, // TODO
+                    "GUILD_MEMBERS_CHUNK" => guild.members_chunk, // TODO
+                    "GUILD_ROLE_CREATE" => guild.role_create GuildRoleCreate: Guild,
+                    "GUILD_ROLE_UPDATE" => guild.role_update GuildRoleUpdate: RoleObject,
+                    "GUILD_ROLE_DELETE" => guild.role_delete, // TODO
+                    "GUILD_SCHEDULED_EVENT_CREATE" => guild.role_scheduled_event_create, // TODO
+                    "GUILD_SCHEDULED_EVENT_UPDATE" => guild.role_scheduled_event_update, // TODO
+                    "GUILD_SCHEDULED_EVENT_DELETE" => guild.role_scheduled_event_delete, // TODO
+                    "GUILD_SCHEDULED_EVENT_USER_ADD" => guild.role_scheduled_event_user_add,
+                    "GUILD_SCHEDULED_EVENT_USER_REMOVE" => guild.role_scheduled_event_user_remove,
+                    "PASSIVE_UPDATE_V1" => guild.passive_update_v1, // TODO
+                    "INTEGRATION_CREATE" => integration.create, // TODO
+                    "INTEGRATION_UPDATE" => integration.update, // TODO
+                    "INTEGRATION_DELETE" => integration.delete, // TODO
+                    "INTERACTION_CREATE" => interaction.create, // TODO
+                    "INVITE_CREATE" => invite.create, // TODO
+                    "INVITE_DELETE" => invite.delete, // TODO
+                    "MESSAGE_CREATE" => message.create,
+                    "MESSAGE_UPDATE" => message.update, // TODO
+                    "MESSAGE_DELETE" => message.delete,
+                    "MESSAGE_DELETE_BULK" => message.delete_bulk,
+                    "MESSAGE_REACTION_ADD" => message.reaction_add, // TODO
+                    "MESSAGE_REACTION_REMOVE" => message.reaction_remove, // TODO
+                    "MESSAGE_REACTION_REMOVE_ALL" => message.reaction_remove_all, // TODO
+                    "MESSAGE_REACTION_REMOVE_EMOJI" => message.reaction_remove_emoji, // TODO
+                    "MESSAGE_ACK" => message.ack,
+                    "PRESENCE_UPDATE" => user.presence_update, // TODO
+                    "RELATIONSHIP_ADD" => relationship.add,
+                    "RELATIONSHIP_REMOVE" => relationship.remove,
+                    "STAGE_INSTANCE_CREATE" => stage_instance.create,
+                    "STAGE_INSTANCE_UPDATE" => stage_instance.update, // TODO
+                    "STAGE_INSTANCE_DELETE" => stage_instance.delete,
+                    "TYPING_START" => user.typing_start,
+                    "USER_UPDATE" => user.update, // TODO
+                    "USER_GUILD_SETTINGS_UPDATE" => user.guild_settings_update,
+                    "VOICE_STATE_UPDATE" => voice.state_update, // TODO
+                    "VOICE_SERVER_UPDATE" => voice.server_update,
+                    "WEBHOOKS_UPDATE" => webhooks.update
+                );
+            }
+            // We received a heartbeat from the server
+            // "Discord may send the app a Heartbeat (opcode 1) event, in which case the app should send a Heartbeat event immediately."
+            GATEWAY_HEARTBEAT => {
+                trace!("GW: Received Heartbeat // Heartbeat Request");
+
+                // Tell the heartbeat handler it should send a heartbeat right away
+
+                let heartbeat_communication = HeartbeatThreadCommunication {
+                    sequence_number: gateway_payload.sequence_number,
+                    op_code: Some(GATEWAY_HEARTBEAT),
+                };
+
+                let heartbeat_thread_communicator = self.get_heartbeat_handler().get_send();
+
+                heartbeat_thread_communicator
+                    .send(heartbeat_communication)
+                    .await
+                    .unwrap();
+            }
+            GATEWAY_RECONNECT => {
+                todo!()
+            }
+            GATEWAY_INVALID_SESSION => {
+                todo!()
+            }
+            // Starts our heartbeat
+            // We should have already handled this in gateway init
+            GATEWAY_HELLO => {
+                warn!("Received hello when it was unexpected");
+            }
+            GATEWAY_HEARTBEAT_ACK => {
+                trace!("GW: Received Heartbeat ACK");
+
+                // Tell the heartbeat handler we received an ack
+
+                let heartbeat_communication = HeartbeatThreadCommunication {
+                    sequence_number: gateway_payload.sequence_number,
+                    op_code: Some(GATEWAY_HEARTBEAT_ACK),
+                };
+
+                let heartbeat_handler = self.get_heartbeat_handler();
+                let heartbeat_thread_communicator = heartbeat_handler.get_send();
+
+                heartbeat_thread_communicator
+                    .send(heartbeat_communication)
+                    .await
+                    .unwrap();
+            }
+            GATEWAY_IDENTIFY
+            | GATEWAY_UPDATE_PRESENCE
+            | GATEWAY_UPDATE_VOICE_STATE
+            | GATEWAY_RESUME
+            | GATEWAY_REQUEST_GUILD_MEMBERS
+            | GATEWAY_CALL_SYNC
+            | GATEWAY_LAZY_REQUEST => {
+                info!(
+                    "Received unexpected opcode ({}) for current state. This might be due to a faulty server implementation and is likely not the fault of chorus.",
+                    gateway_payload.op_code
+                );
+            }
+            _ => {
+                warn!("Received unrecognized gateway op code ({})! Please open an issue on the chorus github so we can implement it", gateway_payload.op_code);
+            }
+        }
+
+        // If we we received a seq number we should let it know
+        if let Some(seq_num) = gateway_payload.sequence_number {
+            let heartbeat_communication = HeartbeatThreadCommunication {
+                sequence_number: Some(seq_num),
+                // Op code is irrelevant here
+                op_code: None,
+            };
+
+            let heartbeat_handler = self.get_heartbeat_handler();
+            let heartbeat_thread_communicator = heartbeat_handler.get_send();
+            heartbeat_thread_communicator
+                .send(heartbeat_communication)
+                .await
+                .unwrap();
         }
     }
+}
 
-    #[tokio::test]
-    async fn test_observer_behavior() {
-        let mut event = GatewayEvent::default();
+#[async_trait(?Send)]
+pub trait GatewayHandleCapable<R, S>
+where
+    R: Stream,
+    S: Sink<Message>,
+{
+    /// Sends json to the gateway with an opcode
+    async fn send_json_event(&self, op_code: u8, to_send: serde_json::Value);
 
-        let new_data = types::GatewayResume {
-            token: "token_3276ha37am3".to_string(),
-            session_id: "89346671230".to_string(),
-            seq: "3".to_string(),
-        };
+    /// Observes an Item `<T: Updateable>`, which will update itself, if new information about this
+    /// item arrives on the corresponding Gateway Thread
+    async fn observe<T: Updateable + Clone + std::fmt::Debug + Composite<T> + Send + Sync>(
+        &self,
+        object: Arc<RwLock<T>>,
+    ) -> Arc<RwLock<T>>;
 
-        let consumer = Arc::new(Consumer {
-            _name: "first".into(),
-            events_received: 0.into(),
-        });
-        event.subscribe(consumer.clone());
-
-        let second_consumer = Arc::new(Consumer {
-            _name: "second".into(),
-            events_received: 0.into(),
-        });
-        event.subscribe(second_consumer.clone());
-
-        event.notify(new_data.clone()).await;
-        event.unsubscribe(&*consumer);
-        event.notify(new_data).await;
-
-        assert_eq!(consumer.events_received.load(Relaxed), 1);
-        assert_eq!(second_consumer.events_received.load(Relaxed), 2);
+    /// Recursively observes and updates all updateable fields on the struct T. Returns an object `T`
+    /// with all of its observable fields being observed.
+    async fn observe_and_into_inner<T: Updateable + Clone + std::fmt::Debug + Composite<T>>(
+        &self,
+        object: Arc<RwLock<T>>,
+    ) -> T {
+        let channel = self.observe(object.clone()).await;
+        let object = channel.read().unwrap().clone();
+        object
     }
+
+    /// Sends an identify event to the gateway
+    async fn send_identify(&self, to_send: types::GatewayIdentifyPayload) {
+        let to_send_value = serde_json::to_value(&to_send).unwrap();
+
+        trace!("GW: Sending Identify..");
+
+        self.send_json_event(GATEWAY_IDENTIFY, to_send_value).await;
+    }
+
+    /// Sends an update presence event to the gateway
+    async fn send_update_presence(&self, to_send: types::UpdatePresence) {
+        let to_send_value = serde_json::to_value(&to_send).unwrap();
+
+        trace!("GW: Sending Update Presence..");
+
+        self.send_json_event(GATEWAY_UPDATE_PRESENCE, to_send_value)
+            .await;
+    }
+
+    /// Sends a resume event to the gateway
+    async fn send_resume(&self, to_send: types::GatewayResume) {
+        let to_send_value = serde_json::to_value(&to_send).unwrap();
+
+        trace!("GW: Sending Resume..");
+
+        self.send_json_event(GATEWAY_RESUME, to_send_value).await;
+    }
+
+    /// Sends a request guild members to the server
+    async fn send_request_guild_members(&self, to_send: types::GatewayRequestGuildMembers) {
+        let to_send_value = serde_json::to_value(&to_send).unwrap();
+
+        trace!("GW: Sending Request Guild Members..");
+
+        self.send_json_event(GATEWAY_REQUEST_GUILD_MEMBERS, to_send_value)
+            .await;
+    }
+
+    /// Sends an update voice state to the server
+    async fn send_update_voice_state(&self, to_send: types::UpdateVoiceState) {
+        let to_send_value = serde_json::to_value(to_send).unwrap();
+
+        trace!("GW: Sending Update Voice State..");
+
+        self.send_json_event(GATEWAY_UPDATE_VOICE_STATE, to_send_value)
+            .await;
+    }
+
+    /// Sends a call sync to the server
+    async fn send_call_sync(&self, to_send: types::CallSync) {
+        let to_send_value = serde_json::to_value(&to_send).unwrap();
+
+        trace!("GW: Sending Call Sync..");
+
+        self.send_json_event(GATEWAY_CALL_SYNC, to_send_value).await;
+    }
+
+    /// Sends a Lazy Request
+    async fn send_lazy_request(&self, to_send: types::LazyRequest) {
+        let to_send_value = serde_json::to_value(&to_send).unwrap();
+
+        trace!("GW: Sending Lazy Request..");
+
+        self.send_json_event(GATEWAY_LAZY_REQUEST, to_send_value)
+            .await;
+    }
+
+    /// Closes the websocket connection and stops all gateway tasks;
+    ///
+    /// Esentially pulls the plug on the gateway, leaving it possible to resume;
+    async fn close(&self);
+}
+
+pub trait HeartbeatHandlerCapable<S: Sink<Message>> {
+    fn new(
+        heartbeat_interval: Duration,
+        websocket_tx: Arc<Mutex<SplitSink<S, Message>>>,
+        kill_rc: tokio::sync::broadcast::Receiver<()>,
+    ) -> Self;
+
+    fn get_send(&self) -> &Sender<HeartbeatThreadCommunication>;
+    fn get_heartbeat_interval(&self) -> Duration;
 }
